+++ Source/Scene/ModelAnimation.js
@@ -40,6 +40,16 @@ function ModelAnimation(options, model, runtimeAnimation) {
   this._loop = defaultValue(options.loop, ModelAnimationLoop.NONE);
 
   /**
+   * If this is defined, it will be used to compute the local animation time
+   * instead of the scene's time.
+   *
+   * @type {ModelAnimation.AnimationTimeCallback}
+   * @default undefined
+   */
+  this.animationTime = options.animationTime;
+  this._prevAnimationTime = undefined;
+
+  /**
    * The event fired when this animation is started.  This can be used, for
    * example, to play a sound or start a particle system, when the animation starts.
    * <p>
@@ -230,4 +240,26 @@ Object.defineProperties(ModelAnimation.prototype, {
     },
   },
 });
+/**
+ * A function used to compute the local animation time for a ModelAnimation.
+ * @callback ModelAnimation.AnimationTimeCallback
+ *
+ * @param {Number} duration The animation's original duration in seconds.
+ * @param {Number} seconds The seconds since the animation started, in scene time.
+ * @returns {Number} Returns the local animation time.
+ *
+ * @example
+ * // Use real time for model animation (also set 
+ * // ModelAnimationCollection#animateWhilePaused)
+ * function animationTime(duration) {
+ *     return Date.now() / 1000 / duration;
+ * }
+ *
+ * @example
+ * // Offset the phase of the animation, so it starts halfway
+ * // through its cycle.
+ * function animationTime(duration, seconds) {
+ *     return seconds / duration + .5;
+ * }
+ */
 export default ModelAnimation;